Added code to propagate the launch sync pipe to ProcessGDBRemote's launch.
This code also is responsible for writing the launch sync byte (currenlty
the value is not important, but it is the process state reported by
the process after the attach).

There is likely some clean-up needed here - I was not seeing the log
lines I expected re: the sync pipe, possible it isn't getting propagated
correctly yet.