disable monitoring of lldb-launched, llgs-attached local debugging on Linux/*BSD.
See https://github.com/tfiala/lldb/issues/50

This is a temporary change to allow Shawn to look at the race behind this
without the extra wait thread.

We may still need to do this, but need to look into zombie prevention.  Essentially
I think we should probably only be waiting on the child pid if llgs either fails
or detaches.  I need to read up on semantics of this in presence of ptrace by another
process.