Remove fork-sync-exec file duplication checking on file open action.
This code was not implemented correctly.  To do this right, the file open
mode would have to be checked and used as part of the key for the "already
opened and dupe" logic.  For now we just undo that portion.

Original check-in:
sha1 96a0b2cdf6b02b8c153f3ab94ca2a42f2e57c676