Added attach handler hijacking to PlatformLinux to match PlatformDarwin.
There are a few parameter differences to one of the calls in Attach.  We
might be able to collapse these into a PlatformPOSIX helper if we parameterize
that (or if I discover that those parameter differences are not needed).

This does not yet resolve the issue, but is related to:
https://github.com/tfiala/lldb/issues/50