Ensure the Process UnixSignals instance gets configured properly.
Adds Host::GetUnixSignals().

Adds a Process constructor variant that takes the UnixSignalsSP for
the process.

Defaults to using Host::GetUnixSignals () when the Process constructor
is not provided with a UnixSignalsSP.  This fixes:
https://github.com/tfiala/lldb/issues/50

since, prior to the change, ProcessGDBRemote for local llgs was
initialized with UnixSignals directly, which is only appropriate
currently for MacOSX.

Modifies ProcessLinux/ProcessFreeBSD to construct Process with the
right UnixSignals, and removes ProcessPOSIX's override of
Process::GetUnixSignals () since they are no longer needed.